Refactor(basic): app.js의 UI 관련 함수를 view.js로 이동
- onUpdateSelectOptions 함수를 view.js로 분리
- doUpdatePricesInCart 함수를 view.js로 분리
- app.js는 각 모듈을 연결하고 초기화하는 역할에 집중
